home

+++ contentReader/src/contentReader/main/Reader.java
@@ -2,5 +2,5 @@ package contentReader.main;
 
 public interface Reader {
 
-	public Long getWordNum();
+	public Integer getWordNum();
 }
